added horovodrun command line option
