+++ bin/horovodrun
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+
+from horovod.run.run import run
+
+if __name__ == '__main__':
+    run()
+
+
